+++ refmt_test/emptyFileComment.t/run.t
@@ -1,3 +1,0 @@
-Format empty file comment
-  $ ../../src/refmt/refmt_impl.exe --print-width 50 ./input.re
-  // file with just a single line comment