+++ refmt_test/singleLineCommentEof.t/run.t
@@ -1,3 +1,0 @@
-Format single line comment at the end of the file
-  $ ../../src/refmt/refmt_impl.exe --print-width 50 ./input.re
-  // let x = 1