+++ test/emptyFileComment.t/run.t
@@ -1,0 +1,3 @@
+Format empty file comment
+  $ refmt ./input.re
+  // file with just a single line comment