+++ test/trailingSpaces.t/run.t
@@ -1,0 +1,9 @@
+Format trailing spaces
+  $ refmt ./input.re
+  /* Copyright (c) 2015-present, Facebook, Inc. All rights reserved. */
+  
+  module M =
+    Something.Create({
+      type resource1 = MyModule.MySubmodule.t;
+      type resource2 = MyModule.MySubmodule.t;
+    });